+++ RxSwift/ObservableType.swift
@@ -7,6 +7,7 @@
 //
 
 /// Represents a push style sequence.
+@dynamicMemberLookup
 public protocol ObservableType: ObservableConvertibleType {
     /**
     Subscribes `observer` to receive events for this sequence.
@@ -43,5 +44,15 @@ extension ObservableType {
         return Observable.create { o in
             return self.subscribe(o)
         }
+    }
+}
+
+/**
+Allows easier mapping by exposing dynamic member keypaths
+So Observable.just("string").map { $0.count } can rewritten as Observable.just("string").count
+*/
+extension ObservableType {
+    public subscript<U>(dynamicMember keyPath: KeyPath<Element, U>) -> Observable<U> {
+        return self.map { $0[keyPath: keyPath]}
     }
 }